feat: efficient reference support
