+++ drive/src/drive/document/insert.rs
@@ -1,6 +1,7 @@
 use grovedb::{Element, TransactionArg};
 use std::collections::HashSet;
 
+use grovedb::reference_path::ReferencePathType::{SiblingReference, UpstreamRootHeightReference};
 use std::option::Option::None;
 
 use crate::contract::document::Document;
@@ -10,9 +11,8 @@ use crate::drive::document::{
     contract_document_type_path,
     contract_documents_keeping_history_primary_key_path_for_document_id,
     contract_documents_keeping_history_primary_key_path_for_document_id_size,
-    contract_documents_keeping_history_storage_time_reference_path,
     contract_documents_keeping_history_storage_time_reference_path_size,
-    contract_documents_primary_key_path,
+    contract_documents_primary_key_path, make_document_reference,
 };
 use crate::drive::flags::StorageFlags;
 use crate::drive::object_size_info::DocumentInfo::{
@@ -142,17 +142,14 @@ impl Drive {
                             document_type.name.as_str(),
                             document.id.as_slice(),
                         );
-                    let contract_storage_path =
-                        contract_documents_keeping_history_storage_time_reference_path(
-                            contract.id.as_bytes(),
-                            document_type.name.as_str(),
-                            document.id.as_slice(),
-                            encoded_time,
-                        );
                     PathFixedSizeKeyElement((
                         document_id_in_primary_path,
                         &[0],
-                        Element::Reference(contract_storage_path, storage_flags.to_element_flags()),
+                        Element::Reference(
+                            SiblingReference(encoded_time),
+                            Some(1),
+                            storage_flags.to_element_flags(),
+                        ),
                     ))
                 } else {
                     let path_max_length =
@@ -509,24 +506,6 @@ impl Drive {
                 // Iteration 2. the index path is now something like Contracts/ContractID/Documents(1)/$ownerId/<ownerId>/toUserId/<ToUserId>/accountReference/<accountReference>
             }
 
-            fn make_document_reference(
-                primary_key_path: [&[u8]; 5],
-                document: &Document,
-                document_type: &DocumentType,
-                storage_flags: &StorageFlags,
-            ) -> Element {
-                // we need to construct the reference to the original document
-                let mut reference_path = primary_key_path
-                    .iter()
-                    .map(|x| x.to_vec())
-                    .collect::<Vec<Vec<u8>>>();
-                reference_path.push(Vec::from(document.id));
-                if document_type.documents_keep_history {
-                    reference_path.push(vec![0]);
-                }
-                Element::Reference(reference_path, storage_flags.to_element_flags())
-            }
-
             // unique indexes will be stored under key "0"
             // non unique indices should have a tree at key "0" that has all elements based off of primary key
             if !index.unique || any_fields_null {
@@ -548,7 +527,6 @@ impl Drive {
                     DocumentAndSerialization((document, _, storage_flags))
                     | DocumentWithoutSerialization((document, storage_flags)) => {
                         let document_reference = make_document_reference(
-                            primary_key_path,
                             document,
                             document_and_contract_info.document_type,
                             storage_flags,
@@ -573,7 +551,6 @@ impl Drive {
                     DocumentAndSerialization((document, _, storage_flags))
                     | DocumentWithoutSerialization((document, storage_flags)) => {
                         let document_reference = make_document_reference(
-                            primary_key_path,
                             document,
                             document_and_contract_info.document_type,
                             storage_flags,
